requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
defusedxml==0.7.1
Django==5.2.1
lxml==5.4.0
numpy==2.2.5
odfpy==1.4.1
pandas==2.2.3
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
striprtf==0.0.29
typing_extensions==4.13.2
tzdata==2025.2
cryptography==42.0.5
